For module_utils, do not look in subdirectories.
If PluginLoader adds subdirectories, then we could end up getting the
wrong file for an import statement.  Imagine if we have a file like
this: module_utils/bar/foo.py and the import line in the module is this:
from ansible.module_utils import foo.  If PluginLoader returns
module_utils/bar/ as one of the paths to search, then AnsiballZ will
include module_utils/bar/foo.py in the zipfile instead of looking for
something else at the toplevel.  The Ansible module will then crash on
the remote machine because it can't find the module it expects at the
location it expects.